# api/openapi.yaml
openapi: 3.1.0
info:
  title: WorkAMA API
  version: 0.1.0
  description: Implemented P0 vertical-slice contract. The full P3 registry remains governed by WorkAMA-Docs/720.
servers:
  - url: http://localhost:20200/api/v1
paths:
...
  /privacy/consents/{policyType}:
    put:
      operationId: setConsentDecision
      security: [{ bearerAuth: [] }]
      tags: [Privacy]
      parameters:
        - { name: policyType, in: path, required: true, schema: { type: string } }
      responses: { '200': { description: Consent decision persisted } }
  /privacy/data-requests:
    get:
      operationId: listDataRequests
      security: [{ bearerAuth: [] }]
      tags: [Privacy]
      responses: { '200': { description: Data request page } }
    post:
      operationId: createDataRequest
      security: [{ bearerAuth: [] }]
      tags: [Privacy]
      responses: { '202': { description: Request accepted } }
...
  /webhooks:
    get:
      operationId: listWebhooks
      security: [{ bearerAuth: [] }]
      tags: [OpenPlatform]
      responses: { '200': { description: Workspace webhooks, secret material omitted, content: { application/json: { schema: { $ref: '#/components/schemas/ItemListWebhook' } } } } }
    post:
      operationId: createWebhook
      security: [{ bearerAuth: [] }]
      tags: [OpenPlatform]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/WebhookCreate' } } }
      responses: { '201': { description: Webhook secret is returned once, content: { application/json: { schema: { $ref: '#/components/schemas/WebhookSecretResponse' } } } } }
  /webhooks/{webhookId}:
    get:
      operationId: getWebhook
      security: [{ bearerAuth: [] }]
      tags: [OpenPlatform]
      parameters: [{ name: webhookId, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: Webhook without secret, content: { application/json: { schema: { $ref: '#/components/schemas/Webhook' } } } } }
    patch:
      operationId: updateWebhook
      security: [{ bearerAuth: [] }]
      tags: [OpenPlatform]
      parameters: [{ name: webhookId, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/WebhookPatch' } } }
      responses: { '200': { description: Updated webhook, content: { application/json: { schema: { $ref: '#/components/schemas/Webhook' } } } } }
    delete:
      operationId: deleteWebhook
      security: [{ bearerAuth: [] }]
      tags: [OpenPlatform]
      parameters: [{ name: webhookId, in: path, required: true, schema: { type: string } }]
      responses: { '204': { description: Webhook revoked } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    WebhookCreate:
      type: object
      additionalProperties: false
      required: [url, events]
      properties:
        url: { type: string, minLength: 1, maxLength: 2048, description: HTTPS endpoint or controlled mock:// or local:// endpoint }
        events: { type: array, minItems: 1, maxItems: 32, items: { type: string } }
        description: { type: string, maxLength: 500, default: '' }
    WebhookPatch:
      type: object
      additionalProperties: false
      properties:
        url: { type: [string, 'null'], maxLength: 2048 }
        events: { type: [array, 'null'], maxItems: 32, items: { type: string } }
        description: { type: [string, 'null'], maxLength: 500 }
        status: { type: [string, 'null'], enum: [active, disabled] }
    Webhook:
      type: object
      required: [id, url, events, description, secret_status, secret_last4, status, failure_count, version, created_at, updated_at]
      properties:
        id: { type: string }
        url: { type: string }
        events: { type: array, items: { type: string } }
        description: { type: string }
        secret_status: { type: string, const: configured }
        secret_last4: { type: string, minLength: 4, maxLength: 4 }
        status: { type: string, enum: [active, disabled, revoked] }
        failure_count: { type: integer, minimum: 0 }
        version: { type: integer, minimum: 1 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WebhookSecretResponse:
      allOf:
        - { $ref: '#/components/schemas/Webhook' }
        - type: object
          required: [secret]
          properties:
            secret: { type: string, minLength: 1, writeOnly: true }
            secret_status: { type: string, const: shown_once }
    ItemListWebhook:
      type: object
      required: [items]
      properties: { items: { type: array, items: { $ref: '#/components/schemas/Webhook' } } }
...